#15358: Add GTest, necessary src file modification, core function in a lib
