+++ tests/tt_metal/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(distributed)
 add_subdirectory(tt_metal)
+add_subdirectory(tools/tt_builder)
