+++ tests/tt_metal/tt_metal/CMakeLists.txt
@@ -85,4 +85,5 @@ add_custom_target(
         unit_tests_llk
         unit_tests_stl
         unit_tests_noc
+        unit_tests_builder
 )
