+++ tt_metal/tools/tt_builder/builder.hpp
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+void build_firmware();
